+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 To see unreleased changes, please see the [CHANGELOG on the main branch guide](https://github.com/gufolabs/gufo_snmp/blob/main/CHANGELOG.md).
 
+## [Unreleased]
+
+### Fixed
+
+* #17: Fix premature end of getbulk.
+
 ## 0.7.0 - 2025-03-05
 
 ### Added